POPSCI-81
https://tracker.nci.nih.gov/browse/POPSCI-81
Also changes to Facets to correct the confusion of biospecimencollection and data collection that was causing an error that caused the information on the explore page to crash.

Popsci-81 changes are changes to the indices so the counts on the stats bar line up with values on the study info pages.